Added in the working_set_bytes metric
Hey Sumo,

Added in a regex match for `working_set_bytes` so that we can keep that metric and send it along to Sumo as well. `working_set_bytes` is used to track the working memory and is what the Kubernetes system uses to trigger OOMkills and restarts against the limits.