Fix additionalLabels in collection-fluent-bit
In https://github.com/SumoLogic/sumologic-kubernetes-collection/pull/690 the prometheus additionalLabels were updated to use `sumologic.com/app` instead of `app`. The entry for `collection-fluent-bit` was missed, causing the helm chart to fail when using the built in prometheus stack due to duplicate labels.

This change should fix that so the built in prometheus stack works with collection-fluent-bit.