fix: logs not coming in after switch from fluentd to otelcol
Given I'm running with Fluentd as logs metadata provider,
When I switch to OtelCol as logs metadata provider in place
(with `helm upgrade` as opposed to `helm uninstall` and `helm install`),
Fluent Bit is not able to pick up the new configuration and keeps
trying to send data to Fluentd service.

To properly fix that, we would need to introduce a breaking change
in Fluent Bit configuration. This should be done in v3.0.
For now, let's disguise OtelCol statefulset as Fluentd by naming
its service with the name of the Fluentd service,
so that Fluent Bit can keep using the Fluentd service name,
even though there's really OtelCol under the hood.

Fixes #1741