fix: metrics not coming in after switch from fluentd to otelcol
Given I'm running with Fluentd as metrics metadata provider
When I switch to OtelCol as metrics metadata provider in place
(with `helm upgrade` as opposed to `helm uninstall` and `helm install`),
Prometheus is not able to pick up the new configuration and keeps
trying to send data to Fluentd service.

Let's disguise OtelCol statefulset as Fluentd by naming
its service with the name of the Fluentd service,
so that Prometheus can keep using the Fluentd service name,
even though there's really OtelCol under the hood.