ci: publish tag versions to the dev repository
This causes a bit of duplication, as the same commit can be published to the
dev repository with two different versions, but it ensures that it's
always possible to derive the dev version from just the repository
state, irrespective of whether a tag is present or not.

This also helps fix an issue in our e2e tests.