fix: increase OTC liveness timeout and period
The current liveness probe doesn't actually check liveness in any
meaningful way - it merely checks if the health_check extension is
functional. We really don't want it to cause restarts under load, which
it can with the current settings. Increase the timeout and period to
mitigate this. We can revisit once
the liveness check is more meaningful.