fix: use custom ServiceMonitor for Prometheus' own metrics
Having the `kube-prometheus-stack` subchart generate the ServiceMonitor to scrape Prometheus' own metrics does not work correctly in a setup with an external Prometheus operator in a different (higher) version.
By "does not work correctly" I mean that the ServiceMonitor points to a Service that does not see any pods and in effect, Prometheus cannot scrape its own metrics.
This is because different versions of Prometheus operator create different labels on Prometheus pods.
To fix this, let's disable the `kube-prometheus-stack` subchart's built-in ServiceMonitor and create our own, just like we do for other collection resources (Fluent Bit, Fluentd, Otelcol).