fix: set cluster field in metadata pipelines
The `cluster` metadata field was set on the collector in Sumo. In the
metadata pipelines, the behaviour was inconsistent. FluentD metrics set
it correctly, but seemingly no other pipeline did. In particular, OT
pipelines were actively deceptive,
pretending to set it in the k8s_tagger processor, which doesn't actually
do anything, and was recently deprecated upstream for this reason.

Additionally, setting the field on the collector actually behaves
subtly differently from sending it as a metric label, which caused one
nasty bug for a client trialing OT.

This change explicitly sets the field based on the configuration value
and removes it from the collector.