Use either minAvailable or maxUnavailable for logs pdb
This change updates the logs pdb.yaml to either use maxUnavailable, if defined,
or minAvailable as a default.

The current behavior is to always include minAvailable since it's specified as
a default in values.yaml file. There's no way in helm to unset this. With this
change it will use maxUnavailable if defined and if not it will fall back to
minAvailable.