tf cleanup: do not modify SUMOLOGIC_BASE_URL
cleanup.sh can complain that SUMOLOGIC_BASE_URL is unbound. It attempts
to modify the URL so that it does not contain a v1.

This seems like some sort of hack that should have been cleaned up. I
don't think we should be doing this, so I am removing references to
SUMOLOGIC_BASE_URL in this script.

Signed-off-by: Eric Chlebek <echlebek@sumologic.com>