chore: upgrade opentelemetry-operator (#3985)
* upgrade opentelemetry-operator to v0.93.0
* add targetallocators CRD
* don't apply yaml linting to CRDs
* Revert "fix: Downgrade helm version in CI (#3984)"

This PR upgrades opentelemetry-operator to v0.93.0. This resolves an incompatibility with helm versions > 3.18.4.

The TargetAllocator type has been moved into its own CRD, so it has been added as well.

As a result of upgrading to opentelemetry-operator v0.93.0, we can revert the workaround we had to add in CI that downgrades helm.

Signed-off-by: Eric Chlebek <echlebek@sumologic.com>