chore: set GODEBUG=netdns=go in Chart.yaml
This commit sets GODEBUG=netdns=go in Chart.yaml, selecting the netgo
DNS resolver instead of the cgo resolver that is made available by the
operating system. Previously, we relied on the collector to be built
with only the netgo resolver. However, future releases of the collector
will include both DNS resolvers, and a known issue in k8s collection
requires the use of the netgo resolver.

Signed-off-by: Eric Chlebek <echlebek@sumologic.com>